fix: removes useEntry wrapper
BREAKING CHANGES:
useEntry is removed and replaced with directly using resource.useEntry
