--- src/use-entry.js
@@ -1,3 +1,0 @@
-const useEntry = (resource, args = []) => resource.useEntry(...args);
-
-export default useEntry;
